+++ examples/nextjs-app-router-todo/src/app/api/test/route.ts
@@ -1,13 +1,13 @@
-import { NextResponse } from "next/server"
-import { remult } from "remult"
-import { Task } from "../../../shared/task"
-import { withRemult } from "../[...remult]/route"
+import { NextResponse } from 'next/server'
+import { remult } from 'remult'
+import { Task } from '../../../shared/task'
+import { api } from '../../../api'
 
 export async function GET(req: Request) {
-  return withRemult(async () => {
+  return api.withRemult(async () => {
     return NextResponse.json({
       result: await remult.repo(Task).count(),
-      user: remult.user
+      user: remult.user,
     })
   })
 }