+++ projects/core/server/remult-api-server.ts
@@ -39,11 +39,10 @@ import type {
 } from '../src/server-action.js'
 import { Action, classBackendMethodsArray } from '../src/server-action.js'
 import { serverActionField } from '../src/server-action-info.js'
-import { initDataProvider } from './initDataProvider.js'
-import { remult } from '../src/remult-proxy.js'
 import { remultStatic } from '../src/remult-static.js'
 import remultAdminHtml from './remult-admin.js'
 import { isOfType } from '../src/isOfType.js'
+import { initDataProviderOrJson } from './initDataProviderOrJson.js'
 
 export interface RemultServerOptions<RequestType> {
   /**Entities to use for the api */
@@ -163,7 +162,8 @@ export function createRemultServerCore<RequestType>(
     options.queueStorage = new InMemoryQueueStorage()
   }
 
-  let dataProvider = initDataProvider(options.dataProvider)
+  let dataProvider = initDataProviderOrJson(options.dataProvider)
+  remultStatic.defaultDataProvider = () => dataProvider
   if (options.ensureSchema === undefined) options.ensureSchema = true
 
   RemultAsyncLocalStorage.enable()
@@ -298,14 +298,19 @@ export class RemultServerImplementation<RequestType>
   }
   withRemultAsync<T>(request: RequestType, what: () => Promise<T>): Promise<T> {
     if (!request) return this.runWithRemult(what)
-    return new Promise<any>((resolve, error) => {
-      return this.withRemult(request, undefined!, () => {
-        try {
-          what().then(resolve).catch(error)
-        } catch (err) {
-          error(err)
-        }
-      })
+
+    return new Promise<any>(async (resolve, error) => {
+      try {
+        return await this.withRemult(request, undefined!, async () => {
+          try {
+            what().then(resolve).catch(error)
+          } catch (err) {
+            error(err)
+          }
+        })
+      } catch (err) {
+        error(err)
+      }
     })
   }
   getEntities(): EntityMetadata<any>[] {
@@ -346,9 +351,13 @@ export class RemultServerImplementation<RequestType>
     throw new Error("Couldn't find entity " + entityKey)
   }
   subscriptionServer: SubscriptionServer
-  withRemult = (req: RequestType, res: GenericResponse, next: VoidFunction) => {
-    this.process(async () => {
-      next()
+  withRemult = async (
+    req: RequestType,
+    res: GenericResponse,
+    next: VoidFunction,
+  ) => {
+    await this.process(async () => {
+      await next()
     })(req, res)
   }
 
@@ -586,18 +595,23 @@ export class RemultServerImplementation<RequestType>
   ) {
     let dataProvider: DataProvider
     if (!options?.skipDataProvider) dataProvider = await this.dataProvider
-    return await withRemult(async (remult) => {
-      var x = remult
-      x.liveQueryPublisher = new LiveQueryPublisher(
-        () => remult.subscriptionServer,
-        () => remult.liveQueryStorage,
-        this.runWithSerializedJsonContextData,
-      )
-      if (!options?.skipDataProvider) x.dataProvider = dataProvider
-      x.subscriptionServer = this.subscriptionServer
-      x.liveQueryStorage = this.liveQueryStorage
-      return await what(x)
-    })
+    return await withRemult(
+      async (remult) => {
+        var x = remult
+        x.liveQueryPublisher = new LiveQueryPublisher(
+          () => remult.subscriptionServer,
+          () => remult.liveQueryStorage,
+          this.runWithSerializedJsonContextData,
+        )
+        if (!options?.skipDataProvider) x.dataProvider = dataProvider
+        x.subscriptionServer = this.subscriptionServer
+        x.liveQueryStorage = this.liveQueryStorage
+        return await what(x)
+      },
+      {
+        dataProvider,
+      },
+    )
   }
 
   process(
@@ -620,23 +634,23 @@ export class RemultServerImplementation<RequestType>
         }
         if (!genReq.params) genReq.params = req['_tempParams']
       }
-      let myReq = new ExpressRequestBridgeToDataApiRequest(genReq)
-      let myRes = new ExpressResponseBridgeToDataApiResponse(
+      let myReq = new RequestBridgeToDataApiRequest(genReq)
+      let myRes = new ResponseBridgeToDataApiResponse(
         origRes,
         req,
         this.options.error,
       )
-      if (remultStatic.asyncContext.isInInitRequest())
-        return await what(
-          remultStatic.asyncContext.getStore()!.remult,
-          myReq,
-          myRes,
-          genReq,
-          origRes,
-          req,
-        )
-      else
-        try {
+      try {
+        if (remultStatic.asyncContext.isInInitRequest())
+          return await what(
+            remultStatic.asyncContext.getStore()!.remult,
+            myReq,
+            myRes,
+            genReq,
+            origRes,
+            req,
+          )
+        else
           await this.runWithRemult(async (remult) => {
             if (req) {
               remultStatic.asyncContext.setInInitRequest(true)
@@ -666,9 +680,10 @@ export class RemultServerImplementation<RequestType>
             }
             await what(remult, myReq, myRes, genReq, origRes, req)
           })
-        } catch (err: any) {
-          myRes.error(err, undefined)
-        }
+      } catch (err: any) {
+        if (origRes) myRes.error(err, undefined)
+        else throw err
+      }
     }
   }
   async getRemult(req: RequestType) {
@@ -1111,14 +1126,14 @@ export class RemultServerImplementation<RequestType>
   }[] = []
 }
 
-class ExpressRequestBridgeToDataApiRequest implements DataApiRequest {
+class RequestBridgeToDataApiRequest implements DataApiRequest {
   get(key: string): any {
     return this.r?.query[key]
   }
 
   constructor(private r: GenericRequestInfo | undefined) {}
 }
-class ExpressResponseBridgeToDataApiResponse implements DataApiResponse {
+class ResponseBridgeToDataApiResponse implements DataApiResponse {
   forbidden(): void {
     this.error({ message: 'Forbidden' }, undefined, 403)
   }